# Resolved conflicts:
#	bin/bridgecommand-bc.exe
#	bin/bridgecommand-ed.exe
#	bin/bridgecommand-ini.exe
#	bin/bridgecommand-mc.exe
#	bin/bridgecommand-mh.exe
#	bin/bridgecommand-rp.exe
#	bin/bridgecommand.exe
#	src/BridgeCommand.cbp

+++ src/OwnShip.cpp
@@ -684,7 +684,7 @@ void OwnShip::load(OwnShipData ownShipData, irr::core::vector3di numberOfContact
     triangleSelectorEnabled = false;
 
     device->getLogger()->log("Own ship points found: ");
-    device->getLogger()->log(irr::core::stringw(contactPoints.size()).c_str());
+    device->getLogger()->log(irr::core::stringw((int)contactPoints.size()).c_str());
 }
 
 void OwnShip::addContactPointFromRay(irr::core::line3d<irr::f32> ray, irr::f32 contactArea)
